Change all predictions to absolute time and render view every second
This allows us to keep rendering regardless of if we have new data.
Even if just one source fails to load, we'll keep counting it down.
ALl times are more precise, because precision was limited to the json refetch rate.

Also, this will allow short term caching in the future where slightly stale data is still good.